feat: add responsive size package, rename type definition files
