+++ package.json
@@ -32,6 +32,7 @@
     "react-native-nitro-modules": "^0.25.2",
     "react-native-ota-hot-update": "^2.2.9",
     "react-native-reanimated": "^3.17.5",
+    "react-native-responsive-fontsize": "^0.5.1",
     "react-native-safe-area-context": "^5.4.0",
     "react-native-screens": "^4.10.0",
     "react-native-svg": "^15.11.2",
