+++ src/components/index.ts
@@ -1,0 +1,1 @@
+export {default as Typography} from '@/components/Typography';
